+++ elevenlabs_rs/src/endpoints/genai/mod.rs
@@ -1,10 +1,12 @@
 pub use super::*;
 pub mod audio_isolation;
 pub mod dubbing;
+pub mod forced_alignment;
 
 pub mod sound_effects;
 pub mod speech_to_text;
 pub mod text_to_dialogue;
 pub mod text_to_voice;
+pub mod tokens;
 pub mod tts;
 pub mod voice_changer;